requirements_craft_ai.txt
# Craft AI Service Requirements
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
aiohttp>=3.9.0
redis>=5.0.0
python-dotenv>=1.0.0

# Provider dependencies
openai>=1.3.0  # For OpenAI provider
groq>=0.4.0    # For Groq provider
requests>=2.31.0  # For Ollama validation

# Optional: Better JSON handling
ujson>=5.8.0

# Optional: Monitoring and logging
structlog>=23.2.0